src/requirements.txt
tqdm
numpy
Pillow
scikit-image
opencv-python
onnxruntime-gpu
